Fixed types in next_pow_2. Added static_casts to ogl/renderer.cpp where appropriate. Fixed DInput and D3D to use the new com_deleter.hpp. Fixed DInput's keyboard.cpp to remember metakeys. Fixed D3D's enumeration.cpp. Added Code::Blocks project files.
git-svn-id: https://spacegameengine.svn.sourceforge.net/svnroot/spacegameengine@156 469afb1b-a615-0410-ae09-9c7fe5ec7362

+++ src/plugins/dinput/src/key_converter.cpp
@@ -151,7 +151,7 @@ sge::dinput::key_converter::key_converter()
 	v[DIK_Z] = KC_Z;
 
 	for(key_vector::size_type i = 0; i < v.size(); ++i)
-		vr[v[i]] = i;
+		vr[v[i]] = static_cast<DWORD>(i);
 }
 
 sge::key_code sge::dinput::key_converter::create_key_code(const DWORD ofs) const
